Consistently use -std=c++20
This project fiddled with CMAKE_CXX_STANDARD throughput the build
generation process, which leads to inconsistent compile and linker
flags. This doesn't work well with newer versions of protobuf which
has polyfills for some C++20 and C++17 features. The use of these
polyfills will lead to linker errors for the consumer of libbraft.a.